// frontend/src/backend.d.ts
export type Time = bigint;
// ...
export interface BulkEmployeeRecord {
    internalId: bigint;
    status: string;
    sickLeave: bigint;
    annualLeave: bigint;
    personalEmail?: string;
    role: EmployeeRole;
    joiningDate: Time;
    businessEmail: string;
    unpaidLeave: bigint;
    casualLeave: bigint;
    employeeId: string;
    department: string;
    phoneNumber: string;
}
// ...
export enum EmployeeRole {
    hrManager = "hrManager",
    manager = "manager",
    superAdmin = "superAdmin",
    employee = "employee",
    hrStaff = "hrStaff"
}